Update patch set 2
Patch Set 2:

It's reportedly working as expected for the ville as well. This should be good to merge in now.

I've collected everything under the single merge commit so you should be able to just pull that one in and get everything.
SHA# 29aa84dfa5698db266224af260b406bc6067853f
https://github.com/Haus1/android_kernel_htc_msm8960/commit/29aa84dfa5698db266224af260b406bc6067853f

android-4.4_wip is the identical minus the huge merge commit

Patch-set: 2